requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
anthropic>=0.25.0
openai>=1.0.0
python-dotenv>=1.0.0
websockets>=12.0
